refactor(identity): bundle hardened runtime core in SDK wheel
